modernize-requirements/evals/fixtures/eval-1-basic-uv/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
pydantic==2.11.1
pydantic-core==2.33.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.3.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
